move retry logic to standalone module
