+++ ydb/_topic_reader/topic_reader.py
@@ -10,7 +10,7 @@ from typing import (
     Callable,
 )
 
-from ..table import RetrySettings
+from .._retries import RetrySettings
 from .._grpc.grpcwrapper.ydb_topic import StreamReadMessage, OffsetsRange
